Summary: massive updates to make portable
Details:
use the settings in the conf/quads.yml file.
There is now a mechanism for having either python
or bash scripts read the yaml config file. This is
done via the yay.sh.  We wrap this up using
bin/load-config.sh

quads should be much more portable now and the only
changes needed are in the conf/quads.yml file.  This
of course will need to be tested and validated.

+++ bin/create-input.sh
@@ -1,5 +1,21 @@
 #!/bin/sh
 
+if [ ! -e $(dirname $0)/load-config.sh ]; then
+    echo "$(basename $0): could not find load-config.sh"
+    exit 1
+fi
+
+source $(dirname $0)/load-config.sh
+
+quads=${quads["install_dir"]}/bin/quads.py
+quads_url=${quads["quads_url"]}
+rt_url=${quads["rt_url"]}
+datadir=${quads["install_dir"]}/data
+exclude_hosts=${quads["exclude_hosts"]}
+domain=${quads["domain"]}
+# define your racks
+racks=${quads["racks"]}
+
 # output might look like ....
 cat > /dev/null <<EOF
    - This is [automatically generated](https://github.com/redhat-performance/ops-tools/tree/master/lab-scheduler).
@@ -12,8 +28,6 @@ cat > /dev/null <<EOF
 | 1 |c10-h30-r630.openstack.example.com | JVKNZ1 | 00:01:AA:B4:A8 | 10.1.2.1 | 10.2.1.1 | [idrac](http://example.com) |01:AA:BC:7D:8A |cloud1 | someuser | [grafana](http://example.com) |
 EOF
 
-# define your racks
-racks="b09 b10 c01 c02 c03 c04 c05 c06 c07 c08 c09 c10"
 
 function print_header() {
     cat <<EOF
@@ -33,36 +47,31 @@ function add_row() {
     uloc=$2
     #echo $arg $uloc
     nodename=$(echo $arg | sed 's/mgmt-//')
-    if [ ! -d /etc/lab/ipmi/$nodename ]; then 
-        mkdir -p /etc/lab/ipmi/$nodename
+    if [ ! -d $datadir/ipmi/$nodename ]; then 
+        mkdir -p $datadir/ipmi/$nodename
     fi
-#    if [ -f /etc/lab/ipmi/$nodename/svctag ]; then
-        svctag=$(cat /etc/lab/ipmi/$nodename/svctag 2>/dev/null)
-#    else
-#        svctag=$(ssh -o PasswordAuthentication=false -o ConnectTimeout=10 $arg racadm getsysinfo 2>/dev/null | egrep "^Service Tag" | awk '{ print $NF }')
-#        echo $svctag > /etc/lab/ipmi/$nodename/svctag
-#    fi
-    if [ -f /etc/lab/ipmi/$nodename/macaddr ]; then
-        macaddr=$(cat /etc/lab/ipmi/$nodename/macaddr)
+    svctag=$(cat $datadir/ipmi/$nodename/svctag 2>/dev/null)
+    if [ -f $datadir/ipmi/$nodename/macaddr ]; then
+        macaddr=$(cat $datadir/ipmi/$nodename/macaddr)
     else
         macaddr=$(hammer host info --name $nodename | grep "MAC:" | awk '{ print $NF }')
-        echo $macaddr > /etc/lab/ipmi/$nodename/macaddr
+        echo $macaddr > $datadir/ipmi/$nodename/macaddr
     fi
     ip=$(host $nodename | awk '{ print $NF }')
     oobip=$(host $arg | awk '{ print $NF }')
     ooburl="<a href=http://$arg/ target=_blank>console</a>"
-    if [ -f /etc/lab/ipmi/$nodename/oobmacaddr ]; then
-        oobmacaddr=$(cat /etc/lab/ipmi/$nodename/oobmacaddr)
+    if [ -f $datadir/ipmi/$nodename/oobmacaddr ]; then
+        oobmacaddr=$(cat $datadir/ipmi/$nodename/oobmacaddr)
     else
         oobmacaddr=$(hammer host info --name $arg | grep "MAC:" | awk '{ print $NF }')
-        echo $oobmacaddr > /etc/lab/ipmi/$nodename/oobmacaddr
+        echo $oobmacaddr > $datadir/ipmi/$nodename/oobmacaddr
     fi
-    workload=$(/root/quads.py --host $nodename)
+    workload=$($quads --host $nodename)
     # need to figure out owner
     if [ "$workload" == "None" ]; then
         owner=""
     else
-        owner=$(/root/quads.py --ls-owner --cloud-only $workload)
+        owner=$($quads --ls-owner --cloud-only $workload)
     fi
     # need to figure out grafana links
     grafana=""
@@ -75,7 +84,7 @@ function find_u() {
 }
 
 TMPHAMMERFILE=$(mktemp /tmp/hammer_host_list_XXXXXXXX)
-hammer host list --per-page 10000 | grep mgmt | egrep -v 'cyclades|s4810|z9000|5548|c08-h05-r930' | awk '{ print $3 }' 1>$TMPHAMMERFILE 2>&1
+hammer host list --per-page 10000 | grep mgmt | egrep -v "$exclude_hosts}" | awk '{ print $3 }' 1>$TMPHAMMERFILE 2>&1
 
 for rack in $racks ; do
     echo "**Rack "$(echo $rack | tr a-z A-Z)"**"
@@ -83,10 +92,6 @@ for rack in $racks ; do
     for h in $(cat $TMPHAMMERFILE | egrep $rack) ; do
         add_row $h $(find_u $h)
     done
-    # add any special hosts that dont conform to naming ...
-    if [ "$rack" == "c08" ]; then
-        add_row mgmt-foreman.example.com 31
-    fi
     echo ""
 
 done
